Add JSON key content for auth (#68)
This change will not force user to deal with file upload for Service Account JSON key and it can be given as String.

Add optional field `json_key_file_content` to get Service Account JSON key as an String input for authentication.

This will add one more option for user to pass Service Account JSON key. With this option the value of Service Account JSON key can be passed as String, instead of file path.